added 2 columns "anycast_server" and "anycast_location".
If the Servername contains a "_", it is assumed that the part before the "_" is the anycast_server name and the part after the "_" is the location of the server f.e. r.ns.at_vie  sets "anycast_server" to "r.ns.at" and "anycast_location" to "vie". The original svr/server fields remain untouched.